Refactor runtime handling and add wrapper context utilities
- Removed redundant functions from `runtime.py` and replaced them with calls to new utility functions in `wrapper_context.py`.
- Introduced `prepareCemdbContext` and related functions to manage CEMDB context preparation.
- Updated tests to reflect changes in function locations and added new tests for CEMDB context handling.
- Enhanced error handling for invalid Docker image references in runtime resolution.
- Added smoke tests for Docker and Apptainer wrappers to ensure proper command execution.

#5
