+++ .github/workflows/ci.yml
@@ -52,10 +52,38 @@ jobs:
           . .venv/bin/activate
           python -m compileall -q src
 
+  smoke:
+    name: Smoke (fake runtimes)
+    runs-on: ubuntu-latest
+    needs: tests
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+
+      - name: Setup uv
+        uses: astral-sh/setup-uv@v6
+
+      - name: Install dependencies
+        run: |
+          uv venv .venv
+          . .venv/bin/activate
+          uv pip install -e ".[dev]"
+
+      - name: Run fake-runtime smoke tests
+        run: |
+          . .venv/bin/activate
+          pytest -q tests/test_smoke_fake_runtimes.py
+
   package:
     name: Package Check
     runs-on: ubuntu-latest
-    needs: tests
+    needs: [tests, smoke]
 
     steps:
       - name: Checkout
