+++ src/kub_cli/runtime.py
@@ -10,7 +10,6 @@ from __future__ import annotations
 from dataclasses import dataclass
 import os
 from pathlib import Path
-import re
 import shutil
 import subprocess
 from typing import Literal, Mapping, Sequence
@@ -20,9 +19,10 @@ from .config import (
     DEFAULT_DOCKER_IMAGE,
     KubConfig,
     SUPPORTED_RUNTIMES,
-    looksLikeContainerReference,
 )
-from .errors import ImageNotFoundError, KubCliError, RunnerNotFoundError, RuntimeSelectionError
+from .errors import KubCliError, RunnerNotFoundError, RuntimeSelectionError
+from .image_resolution import resolveApptainerExecutionImage, resolveDockerExecutionImage
+from . import image_resolution as imageResolution
 from .logging_utils import LOGGER, formatCommand
 
 
@@ -40,48 +40,9 @@ class RuntimeResolution:
 
 
 def deriveApptainerOrasReference(dockerImageReference: str) -> str:
-    """Derive Apptainer ORAS source from a Docker image reference.
+    """Backward-compatible export for ORAS derivation."""
 
-    Example:
-    `ghcr.io/org/app:master` -> `oras://ghcr.io/org/app:master-sif`
-    """
-
-    normalized = dockerImageReference.strip()
-    if not normalized:
-        raise KubCliError("Docker image reference cannot be empty.")
-
-    if "@" in normalized:
-        raise KubCliError(
-            "Cannot derive Apptainer ORAS reference from digest-based Docker image. "
-            "Provide a tag-based Docker image reference."
-        )
-
-    if normalized.startswith("oras://"):
-        return normalized
-
-    if "://" in normalized:
-        raise KubCliError(
-            "Docker image reference for ORAS derivation must not include a URI scheme. "
-            "Use format like ghcr.io/org/image:tag"
-        )
-
-    lastSlash = normalized.rfind("/")
-    lastColon = normalized.rfind(":")
-
-    if lastColon > lastSlash:
-        repository = normalized[:lastColon]
-        tag = normalized[lastColon + 1 :]
-    else:
-        repository = normalized
-        tag = "latest"
-
-    if not repository or not tag:
-        raise KubCliError(
-            "Invalid Docker image reference for ORAS derivation. "
-            "Expected format like ghcr.io/org/image:tag"
-        )
-
-    return f"oras://{repository}:{tag}-sif"
+    return imageResolution.deriveApptainerOrasReference(dockerImageReference)
 
 
 def getRuntimeCandidateImage(config: KubConfig, runtime: ResolvedRuntime) -> str | None:
@@ -185,188 +146,6 @@ def tryResolveRunnerExecutable(runnerValue: str) -> str | None:
     return shutil.which(normalized)
 
 
-def resolveApptainerExecutionImage(config: KubConfig) -> str:
-    """Resolve Apptainer execution image reference (local path or oras:// URI)."""
-
-    explicitImage = resolveExplicitApptainerImage(config)
-    if explicitImage is not None:
-        return explicitImage
-
-    localDefaultImage = resolveLocalDefaultApptainerImage(config)
-    if localDefaultImage is not None:
-        return localDefaultImage
-
-    dockerReference = resolveDockerReferenceForApptainerDerivation(config)
-    return deriveApptainerOrasReference(dockerReference)
-
-
-def resolveExplicitApptainerImage(config: KubConfig) -> str | None:
-    candidates = [config.imageOverride, config.imageApptainer, config.image]
-
-    for candidate in candidates:
-        if candidate is None:
-            continue
-
-        normalizedReference = candidate.strip()
-        if not normalizedReference:
-            continue
-
-        return normalizeApptainerImageReference(normalizedReference)
-
-    return None
-
-
-def normalizeApptainerImageReference(reference: str) -> str:
-    normalizedReference = reference.strip()
-
-    if normalizedReference.startswith("docker://"):
-        raise ImageNotFoundError(
-            "Apptainer image reference must use oras:// (or a local .sif path), "
-            "not docker://."
-        )
-
-    if normalizedReference.startswith("oras://"):
-        return normalizedReference
-
-    if "://" in normalizedReference:
-        raise ImageNotFoundError(
-            "Unsupported Apptainer image URI scheme. "
-            "Use oras://<registry>/<image>:<tag>-sif or a local .sif path."
-        )
-
-    if looksLikeContainerReference(normalizedReference):
-        return f"oras://{normalizedReference}"
-
-    imagePath = Path(normalizedReference).expanduser()
-
-    if not imagePath.exists():
-        raise ImageNotFoundError(f"Container image not found: '{imagePath}'.")
-
-    if imagePath.is_dir():
-        raise ImageNotFoundError(
-            f"Container image must be a file, got directory: '{imagePath}'."
-        )
-
-    return str(imagePath)
-
-
-def resolveLocalDefaultApptainerImage(config: KubConfig) -> str | None:
-    candidateFilenames = [
-        deriveDefaultApptainerImageFilename(config),
-        deriveLegacyDefaultApptainerImageFilename(config),
-    ]
-
-    for filename in candidateFilenames:
-        candidatePath = (Path.cwd() / filename).resolve()
-        if not candidatePath.exists():
-            continue
-
-        if candidatePath.is_dir():
-            raise ImageNotFoundError(
-                f"Container image must be a file, got directory: '{candidatePath}'."
-            )
-
-        return str(candidatePath)
-
-    return None
-
-
-def resolveDockerReferenceForApptainerDerivation(config: KubConfig) -> str:
-    candidates = [config.imageDocker, config.image, DEFAULT_DOCKER_IMAGE]
-
-    for candidate in candidates:
-        if candidate is None:
-            continue
-
-        normalized = candidate.strip()
-        if not normalized:
-            continue
-
-        if normalized.startswith("docker://"):
-            normalized = normalized[len("docker://") :]
-
-        if "://" in normalized:
-            continue
-
-        if looksLikeContainerReference(normalized):
-            return normalized
-
-    return DEFAULT_DOCKER_IMAGE
-
-
-def deriveDefaultApptainerImageFilename(config: KubConfig) -> str:
-    dockerReference = resolveDockerReferenceForApptainerDerivation(config)
-    _, tag = splitImageReference(dockerReference)
-
-    normalizedTag = tag
-    if normalizedTag.endswith("-sif"):
-        normalizedTag = normalizedTag[: -len("-sif")]
-    if normalizedTag.endswith(".sif"):
-        normalizedTag = normalizedTag[: -len(".sif")]
-    if not normalizedTag:
-        normalizedTag = "latest"
-
-    return f"kub-{sanitizePathToken(normalizedTag)}.sif"
-
-
-def deriveLegacyDefaultApptainerImageFilename(config: KubConfig) -> str:
-    dockerReference = resolveDockerReferenceForApptainerDerivation(config)
-    repository, tag = splitImageReference(dockerReference)
-    imageName = repository.rsplit("/", maxsplit=1)[-1] or "kub-image"
-
-    normalizedTag = tag
-    if normalizedTag.endswith("-sif"):
-        normalizedTag = normalizedTag[: -len("-sif")]
-    if normalizedTag.endswith(".sif"):
-        normalizedTag = normalizedTag[: -len(".sif")]
-    if not normalizedTag:
-        normalizedTag = "latest"
-
-    return f"{sanitizePathToken(imageName)}-{sanitizePathToken(normalizedTag)}.sif"
-
-
-def splitImageReference(reference: str) -> tuple[str, str]:
-    normalized = reference.strip()
-    if not normalized:
-        return "kub-image", "latest"
-
-    withoutDigest = normalized.split("@", maxsplit=1)[0]
-    lastSlash = withoutDigest.rfind("/")
-    lastColon = withoutDigest.rfind(":")
-
-    if lastColon > lastSlash:
-        repository = withoutDigest[:lastColon]
-        tag = withoutDigest[lastColon + 1 :]
-    else:
-        repository = withoutDigest
-        tag = "latest"
-
-    if not repository:
-        repository = "kub-image"
-    if not tag:
-        tag = "latest"
-
-    return repository, tag
-
-
-def sanitizePathToken(value: str) -> str:
-    sanitized = re.sub(r"[^A-Za-z0-9._-]+", "-", value).strip("-")
-    return sanitized or "image"
-
-
-def resolveDockerExecutionImage(config: KubConfig) -> str:
-    """Resolve Docker image reference used at runtime execution."""
-
-    imageReference = getRuntimeCandidateImage(config, "docker")
-    if imageReference is None:
-        raise ImageNotFoundError(
-            "No Docker image configured for runtime 'docker'. "
-            "Set --image, KUB_IMAGE_DOCKER, or KUB_IMAGE."
-        )
-
-    return imageReference
-
-
 def resolveRuntimeForExecution(config: KubConfig) -> RuntimeResolution:
     """Resolve runtime backend, executable, and image for application execution."""
 
@@ -447,14 +226,23 @@ def resolveAutoRuntime(config: KubConfig) -> RuntimeResolution:
         dockerRunnerValue = getRunnerValue(config, "docker")
         dockerRunner = tryResolveRunnerExecutable(dockerRunnerValue)
         if dockerRunner is not None:
-            return RuntimeResolution(
-                runtime="docker",
-                runnerPath=dockerRunner,
-                imageReference=dockerImage,
+            try:
+                imageReference = resolveDockerExecutionImage(
+                    config,
+                    strictImageOverride=False,
+                    strictLegacyImage=False,
+                )
+                return RuntimeResolution(
+                    runtime="docker",
+                    runnerPath=dockerRunner,
+                    imageReference=imageReference,
+                )
+            except KubCliError as error:
+                diagnostics.append(f"Docker not selected: {error}")
+        else:
+            diagnostics.append(
+                "Docker not selected: runner not available in PATH or not executable."
             )
-        diagnostics.append(
-            "Docker not selected: runner not available in PATH or not executable."
-        )
     else:
         diagnostics.append(
             "Docker not selected: no Docker image configured."
@@ -539,9 +327,14 @@ class DockerCommandBuilder:
     def resolveImage(self) -> str:
         return resolveDockerExecutionImage(self.config)
 
-    def build(self, forwardedArgs: Sequence[str]) -> list[str]:
+    def build(
+        self,
+        forwardedArgs: Sequence[str],
+        *,
+        imageReference: str | None = None,
+    ) -> list[str]:
         runner = self.resolveRunner()
-        imageReference = self.resolveImage()
+        resolvedImageReference = imageReference or self.resolveImage()
 
         command: list[str] = [runner, "run", "--rm"]
 
@@ -567,7 +360,7 @@ class DockerCommandBuilder:
         for key, value in self.config.env.items():
             command.extend(["--env", f"{key}={value}"])
 
-        command.append(imageReference)
+        command.append(resolvedImageReference)
         command.append(self.appName)
         command.extend(forwardedArgs)
 
@@ -615,23 +408,27 @@ class KubAppRunner:
         if runtimeResolution.runtime == "apptainer":
             builder = ApptainerCommandBuilder(appName=appName, config=self.config)
             command = builder.build(forwardedArgs)
-            try:
-                if shouldUseApptainerExecForLocalImage(
-                    runnerPath=runtimeResolution.runnerPath,
-                    imageReference=runtimeResolution.imageReference,
-                    appName=appName,
-                ):
-                    command = builder.buildExec(forwardedArgs)
-                    if self.config.verbose:
-                        LOGGER.debug(
-                            "Apptainer app '%s' not found in local image; using exec fallback.",
-                            appName,
-                        )
-            except KeyboardInterrupt:
-                return 130
+            if not dryRun:
+                try:
+                    if shouldUseApptainerExecForLocalImage(
+                        runnerPath=runtimeResolution.runnerPath,
+                        imageReference=runtimeResolution.imageReference,
+                        appName=appName,
+                    ):
+                        command = builder.buildExec(forwardedArgs)
+                        if self.config.verbose:
+                            LOGGER.debug(
+                                "Apptainer app '%s' not found in local image; using exec fallback.",
+                                appName,
+                            )
+                except KeyboardInterrupt:
+                    return 130
         else:
             builder = DockerCommandBuilder(appName=appName, config=self.config)
-            command = builder.build(forwardedArgs)
+            command = builder.build(
+                forwardedArgs,
+                imageReference=runtimeResolution.imageReference,
+            )
 
         if self.config.verbose:
             LOGGER.debug(
