+++ tests/test_img_cli.py
@@ -209,3 +209,51 @@ def testApptainerPullRejectsDockerSource(
 
     assert result.exit_code == 2
     assert "must use oras://" in result.output
+
+
+def testPathAutoResolvesToDockerWhenApptainerUnavailable(
+    cliRunner: CliRunner,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    def fakeWhich(name: str) -> str | None:
+        if name == "docker":
+            return "/usr/bin/docker"
+        return None
+
+    monkeypatch.setattr("kub_cli.runtime.shutil.which", fakeWhich)
+
+    result = cliRunner.invoke(
+        imgApp,
+        [
+            "path",
+            "--runtime",
+            "auto",
+        ],
+    )
+
+    assert result.exit_code == 0
+    assert result.output.strip() == "ghcr.io/feelpp/ktirio-urban-building:master"
+
+
+def testAppsAutoRejectsWhenRuntimeResolvesToDocker(
+    cliRunner: CliRunner,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    def fakeWhich(name: str) -> str | None:
+        if name == "docker":
+            return "/usr/bin/docker"
+        return None
+
+    monkeypatch.setattr("kub_cli.runtime.shutil.which", fakeWhich)
+
+    result = cliRunner.invoke(
+        imgApp,
+        [
+            "apps",
+            "--runtime",
+            "auto",
+        ],
+    )
+
+    assert result.exit_code == 2
+    assert "only available with Apptainer runtime" in result.output
